@@ -10,25 +10,27 @@
     "install:clean": "rm -rf node_modules/ && rm -rf package-lock.json && rm -rf .next/ && npm install && npm run dev"
   },
   "dependencies": {
-    "@material-ui/core": "4.11.3",
-    "@material-ui/icons": "4.11.2",
-    "@prisma/client": "^3.12.0",
-    "@zeit/next-css": "1.0.1",
-    "@zeit/next-sass": "1.0.1",
+
+    "@emotion/react": "^11.9.0",
+    "@emotion/styled": "^11.8.1",
+    "@mui/icons-material": "^5.6.1",
+    "@mui/material": "^5.6.1",
+    "@mui/styles": "^5.6.1",
+
     "chartist": "0.10.1",
     "classnames": "2.2.6",
     "history": "5.0.0",
-    "next": "10.0.5",
+    "next": "^12.1.5",
     "next-compose-plugins": "2.2.1",
     "next-images": "1.3.1",
     "path": "0.12.7",
     "perfect-scrollbar": "1.5.0",
     "prisma": "^3.12.0",
     "prop-types": "15.7.2",
-    "react": "17.0.1",
+    "react": "17.0.2",
     "react-chartist": "0.14.3",
-    "react-dom": "17.0.1",
-    "react-swipeable-views": "0.13.9"
+    "react-dom": "17.0.2",
+    "react-swipeable-views": "0.14.0"
   },
   "eslintConfig": {
     "extends": "react-app"
@@ -46,7 +48,9 @@
     ]
   },
   "optionalDependencies": {
+
     "node-sass": "^7.0.1",
+
     "typescript": "4.1.3"
   }
 }
