@@ -3,9 +3,8 @@ import React from "react";
 import classNames from "classnames";
 // nodejs library to set properties for components
 import PropTypes from "prop-types";
-// @material-ui/core components
-import { makeStyles } from "@material-ui/core/styles";
-// @material-ui/icons
+import makeStyles from '@mui/styles/makeStyles';
+// @mui/icons-material
 
 // core components
 import styles from "assets/jss/nextjs-material-dashboard/components/cardHeaderStyle.js";
