@@ -1,14 +1,13 @@
 import React from "react";
 import classNames from "classnames";
 import PropTypes from "prop-types";
-// @material-ui/core components
-import { makeStyles } from "@material-ui/core/styles";
-import FormControl from "@material-ui/core/FormControl";
-import InputLabel from "@material-ui/core/InputLabel";
-import Input from "@material-ui/core/Input";
-// @material-ui/icons
-import Clear from "@material-ui/icons/Clear";
-import Check from "@material-ui/icons/Check";
+import makeStyles from '@mui/styles/makeStyles';
+import FormControl from "@mui/material/FormControl";
+import InputLabel from "@mui/material/InputLabel";
+import Input from "@mui/material/Input";
+// @mui/icons-material
+import Clear from "@mui/icons-material/Clear";
+import Check from "@mui/icons-material/Check";
 // core components
 import styles from "assets/jss/nextjs-material-dashboard/components/customInputStyle.js";
 
