@@ -1,10 +1,9 @@
 /*eslint-disable*/
 import React from "react";
 import PropTypes from "prop-types";
-// @material-ui/core components
-import { makeStyles } from "@material-ui/core/styles";
-import ListItem from "@material-ui/core/ListItem";
-import List from "@material-ui/core/List";
+import makeStyles from '@mui/styles/makeStyles';
+import ListItem from "@mui/material/ListItem";
+import List from "@mui/material/List";
 // core components
 import styles from "assets/jss/nextjs-material-dashboard/components/footerStyle.js";
 
