@@ -2,14 +2,13 @@ import React from "react";
 import classNames from "classnames";
 import PropTypes from "prop-types";
 import { useRouter } from "next/router";
-// @material-ui/core components
-import { makeStyles } from "@material-ui/core/styles";
-import AppBar from "@material-ui/core/AppBar";
-import Toolbar from "@material-ui/core/Toolbar";
-import IconButton from "@material-ui/core/IconButton";
-import Hidden from "@material-ui/core/Hidden";
-// @material-ui/icons
-import Menu from "@material-ui/icons/Menu";
+import makeStyles from '@mui/styles/makeStyles';
+import AppBar from "@mui/material/AppBar";
+import Toolbar from "@mui/material/Toolbar";
+import IconButton from "@mui/material/IconButton";
+import Hidden from "@mui/material/Hidden";
+// @mui/icons-material
+import Menu from "@mui/icons-material/Menu";
 // core components
 import AdminNavbarLinks from "./AdminNavbarLinks.js";
 import RTLNavbarLinks from "./RTLNavbarLinks.js";
@@ -46,7 +45,7 @@ export default function Header(props) {
             {makeBrand()}
           </Button>
         </div>
-        <Hidden smDown implementation="css">
+        <Hidden mdDown implementation="css">
           {props.rtlActive ? <RTLNavbarLinks /> : <AdminNavbarLinks />}
         </Hidden>
         <Hidden mdUp implementation="css">
@@ -54,7 +53,7 @@ export default function Header(props) {
             color="inherit"
             aria-label="open drawer"
             onClick={props.handleDrawerToggle}
-          >
+            size="large">
             <Menu />
           </IconButton>
         </Hidden>
