@@ -1,12 +1,11 @@
 import React from "react";
 import classNames from "classnames";
 import PropTypes from "prop-types";
-// @material-ui/core components
-import { makeStyles } from "@material-ui/core/styles";
-import Snack from "@material-ui/core/Snackbar";
-import IconButton from "@material-ui/core/IconButton";
-// @material-ui/icons
-import Close from "@material-ui/icons/Close";
+import makeStyles from '@mui/styles/makeStyles';
+import Snack from "@mui/material/Snackbar";
+import IconButton from "@mui/material/IconButton";
+// @mui/icons-material
+import Close from "@mui/icons-material/Close";
 // core components
 import styles from "assets/jss/nextjs-material-dashboard/components/snackbarContentStyle.js";
 
@@ -26,7 +25,7 @@ export default function Snackbar(props) {
         aria-label="Close"
         color="inherit"
         onClick={() => props.closeNotification()}
-      >
+        size="large">
         <Close className={classes.close} />
       </IconButton>,
     ];
