@@ -1,7 +1,6 @@
 import React from "react";
 import PropTypes from "prop-types";
-// @material-ui/core components
-import { makeStyles } from "@material-ui/core/styles";
+import makeStyles from '@mui/styles/makeStyles';
 // core components
 import styles from "assets/jss/nextjs-material-dashboard/components/typographyStyle.js";
 
