@@ -1,45 +1,38 @@
 import React from "react";
-import Document, { Head, Main, NextScript } from "next/document";
-import { ServerStyleSheets } from "@material-ui/styles";
+import Document, { Html, Head, Main, NextScript } from "next/document";
+import { ServerStyleSheets } from "@mui/styles";
 
 class MyDocument extends Document {
   render() {
     return (
-      <html lang="en">
+      <Html lang="en">
         <Head>
           <meta charSet="utf-8" />
           <meta name="theme-color" content="#000000" />
-          <link rel="shortcut icon" href={require("assets/img/favicon.png")} />
+          <link
+            rel="shortcut icon"
+            //  href={require("assets/img/favicon.png")}
+            href={"/assets/img/favicon.png"}
+          />
           <link
             rel="apple-touch-icon"
             sizes="76x76"
-            href={require("assets/img/apple-icon.png")}
+            // href={require("/assets/img/apple-icon.png")}
+            href={"/assets/img/apple-icon.png"}
           />
           {/* Fonts and icons */}
-          <link
-            rel="stylesheet"
-            href="//cdn.jsdelivr.net/chartist.js/latest/chartist.min.css"
-          />
+          <link rel="stylesheet" href="//cdn.jsdelivr.net/chartist.js/latest/chartist.min.css" />
           <script src="//cdn.jsdelivr.net/chartist.js/latest/chartist.min.js"></script>
-          <link
-            rel="stylesheet"
-            href="https://fonts.googleapis.com/css?family=Roboto:300,400,500,700|Material+Icons"
-          />
-          <link
-            href="https://fonts.googleapis.com/icon?family=Material+Icons"
-            rel="stylesheet"
-          />
-          <link
-            rel="stylesheet"
-            href="//maxcdn.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css"
-          />
+          <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Roboto:300,400,500,700|Material+Icons" />
+          <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet" />
+          <link rel="stylesheet" href="//maxcdn.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css" />
         </Head>
         <body>
           <div id="page-transition"></div>
           <Main />
           <NextScript />
         </body>
-      </html>
+      </Html>
     );
   }
 }
