@@ -1,6 +1,5 @@
 import React from "react";
-// @material-ui/core components
-import { makeStyles } from "@material-ui/core/styles";
+import makeStyles from '@mui/styles/makeStyles';
 // layout for this page
 import Admin from "layouts/Admin.js";
 // core components
