@@ -1,7 +1,6 @@
 import React from "react";
-// @material-ui/core components
-import { makeStyles } from "@material-ui/core/styles";
-import InputLabel from "@material-ui/core/InputLabel";
+import makeStyles from '@mui/styles/makeStyles';
+import InputLabel from "@mui/material/InputLabel";
 // layout for this page
 import Admin from "layouts/Admin.js";
 // core components
@@ -15,7 +14,7 @@ import CardAvatar from "components/Card/CardAvatar.js";
 import CardBody from "components/Card/CardBody.js";
 import CardFooter from "components/Card/CardFooter.js";
 
-import avatar from "assets/img/faces/marc.jpg";
+const avatar = "/assets/img/faces/marc.jpg";
 
 const styles = {
   cardCategoryWhite: {
