@@ -12,7 +12,7 @@ const headerLinksStyle = (theme) => ({
     "& > div": {
       marginTop: "0",
     },
-    [theme.breakpoints.down("sm")]: {
+    [theme.breakpoints.down('md')]: {
       margin: "10px 15px !important",
       float: "none !important",
       paddingTop: "1px",
@@ -32,7 +32,7 @@ const headerLinksStyle = (theme) => ({
     margin: "0px",
   },
   buttonLink: {
-    [theme.breakpoints.down("sm")]: {
+    [theme.breakpoints.down('md')]: {
       display: "flex",
       margin: "10px 15px 0",
       width: "-webkit-fill-available",
@@ -57,7 +57,7 @@ const headerLinksStyle = (theme) => ({
     },
   },
   searchButton: {
-    [theme.breakpoints.down("sm")]: {
+    [theme.breakpoints.down('md')]: {
       top: "-50px !important",
       marginRight: "22px",
       float: "right",
@@ -89,20 +89,20 @@ const headerLinksStyle = (theme) => ({
       verticalAlign: "middle",
       display: "block",
     },
-    [theme.breakpoints.down("sm")]: {
+    [theme.breakpoints.down('md')]: {
       ...defaultFont,
       fontSize: "14px",
       marginRight: "8px",
     },
   },
   manager: {
-    [theme.breakpoints.down("sm")]: {
+    [theme.breakpoints.down('md')]: {
       width: "100%",
     },
     display: "inline-block",
   },
   searchWrapper: {
-    [theme.breakpoints.down("sm")]: {
+    [theme.breakpoints.down('md')]: {
       width: "-webkit-fill-available",
       margin: "10px 15px 0",
     },
