@@ -30,7 +30,7 @@ const sidebarStyle = (theme) => ({
       position: "fixed",
       height: "100%",
     },
-    [theme.breakpoints.down("sm")]: {
+    [theme.breakpoints.down('md')]: {
       width: drawerWidth,
       ...boxShadow,
       position: "fixed",
@@ -55,7 +55,7 @@ const sidebarStyle = (theme) => ({
       left: "auto !important",
       right: "0 !important",
     },
-    [theme.breakpoints.down("sm")]: {
+    [theme.breakpoints.down('md')]: {
       left: "0  !important",
       right: "auto !important",
     },
