@@ -5,12 +5,18 @@ import PropTypes from "prop-types";
 // nodejs library that concatenates classes
 import classnames from "classnames";
 
-import imagine1 from "assets/img/sidebar-1.jpg";
-import imagine2 from "assets/img/sidebar-2.jpg";
-import imagine3 from "assets/img/sidebar-3.jpg";
-import imagine4 from "assets/img/sidebar-4.jpg";
 
 import Button from "components/CustomButtons/Button.js";
+
+// import imagine1 from "assets/img/sidebar-1.jpg";
+// import imagine2 from "assets/img/sidebar-2.jpg";
+// import imagine3 from "assets/img/sidebar-3.jpg";
+// import imagine4 from "assets/img/sidebar-4.jpg";
+
+const imagine2 = "assets/img/sidebar-2.jpg";
+const imagine1 = "assets/img/sidebar-1.jpg";
+const imagine3 = "assets/img/sidebar-3.jpg";
+const imagine4 = "assets/img/sidebar-4.jpg";
 
 export default function FixedPlugin(props) {
   const [classes, setClasses] = React.useState("dropdown show");
