@@ -1,11 +1,18 @@
 const withPlugins = require("next-compose-plugins");
 const withImages = require("next-images");
-const withSass = require("@zeit/next-sass");
-const withCSS = require("@zeit/next-css");
+// const withSass = require("@zeit/next-sass");
+// const withCSS = require("@zeit/next-css");
 const webpack = require("webpack");
 const path = require("path");
 
-module.exports = withPlugins([[withSass], [withImages], [withCSS]], {
+// module.exports = withPlugins([[withSass], [withImages], [withCSS]], {
+//   webpack(config, options) {
+//     config.resolve.modules.push(path.resolve("./"));
+//     return config;
+//   },
+// });
+
+module.exports = withPlugins([[withImages]], {
   webpack(config, options) {
     config.resolve.modules.push(path.resolve("./"));
     return config;
